fix(layout): 각주 모양 보존·개체 판정 범위·공개 식별자를 리뷰대로 바로잡는다
PR 리뷰 세 건을 합성 문서로 전부 재현한 뒤 고쳤다. 실물 코퍼스(헌법주석 1,030쪽)
렌더는 각주 블록 3,564개가 **좌표까지 완전히 동일**해, 셋 다 잠재 결함만 닫는다.

**처음 잰 구역의 각주 모양을 이월 입력에 실어 보낸다.** 번호 라벨은 구역 각주 모양의
장식 문자·번호 모양을 따르므로, 이어지는 조각을 다음 쪽 구역의 모양으로 다시 만들면
라벨 길이가 달라져 조판 문자열이 통째로 밀린다 — 폭이 같아도 `placedLength`가 다른
글자를 가리켜 `(1)`에서 `1`로 바뀐 재현에서 이어지는 단어가 두 글자 잘렸다. 예약도 그
모양으로 쟀으므로 같은 값을 써야 예약 ≡ 배치가 성립한다(`sizeResolver`와 같은 규약,
R44 #1). `measure`가 처음 잰 모양을 `Input.footnoteShape`에 각인해 나르고, 구분선
여백·길이는 쪽의 몫이라 그대로 현재 구역 모양을 쓴다.

**개체 판정의 범위를 분할 금지와 같은 각주 단위로 맞춘다.** 분할 금지는 각주 전체에
적용하는데 CT 높이 보존은 개체가 직접 붙은 문단만 봤다. 개체가 뒤 문단에만 있으면 앞
문단만 캐시 합으로 줄어(64 → 46.88pt) 그 문단의 마지막 글줄 위로 뒤 문단의 그림이
올라온다. 이제 `measure`가 `noteId`별로 한 번 판정해 그 각주의 전 문단에 적용하고,
**예약도 같은 범위를 본다** — 개체를 담은 각주의 문단은 그 문단에 개체가 없어도 캐시
빠른 길로 가지 않는다(가면 예약이 캐시 합, 배치가 CT 높이로 갈린다).

**공개 API에서도 각주 식별자를 지정할 수 있다.** 공개 `Input.init`이 식별자를 표시
번호로 고정해, 반환된 `overflow`의 이월 조각과 다음 쪽의 새 각주가 같은 번호일 때
구분할 방법이 없었다 — 둘이 한 그룹으로 묶여, 남은 자리에 들어가는 이월 조각까지 함께
다음 쪽으로 밀린다. `noteId:` 인자를 더하고 생략 시 종전대로 표시 번호를 쓴다.

파일 길이 상한(SwiftLint error)에 닿아 각주 문단 측정(`HwpFootnoteNoteMeasurement`)과
예약 측정(`HwpFootnoteCoordinatorMeasurement`)을 각자 파일로 갈랐다. 그 확장이 쓰는
저장 프로퍼티만 모듈 범위로 열었다(공개 표면 불변).

검증: 새 재현 3종(모양 변경 문자 유실·뒤 문단 개체의 앞 문단 높이·표시 번호 병합)이
수정 전 실패·수정 후 통과. 전체 `swift test` 4번들 통과(CoreHwp 1,622·HwpKitCore 845·
Native 281·HwpKit 192). 코퍼스 재렌더는 쪽수 1,030·겹침 0쪽·한글 PDF 대조
1,027/1,030으로 수정 전과 같고, 각주 블록 diff가 0이다.
