+++ Sources/HwpKitCore/AGENTS.md
@@ -1272,7 +1272,7 @@ paraShape와 같은 값**이어야 한다.
   (`HwpFootnoteLayout.separatorLine`, 흐름 모드도 같은 산식). ③ 각주 사이 여백은 앞
   각주 마지막 줄 상자에서 다음 첫 줄까지다 — 마지막 줄의 줄 간격을 **대체**한다. "같은
   각주"의 판정은 표시 번호가 아니라 **식별자**(`Input.noteId`, 리셋 없는 문서 순서
-  일련번호)다: 쪽마다 번호를 새로 시작하면 (표 134 모드 2) 이월된 각주와 그 쪽의 새
+  일련번호, 공개 `Input.init`은 `noteId:` 인자로 지정하고 생략하면 표시 번호를 쓴다)다: 쪽마다 번호를 새로 시작하면 (표 134 모드 2) 이월된 각주와 그 쪽의 새
   각주가 같은 번호를 갖는데, 번호로 묶으면 둘이 한 각주가 돼 사이 여백이 사라지고 앞
   조각의 마지막 줄 간격이 높이에 남으며 하나가 안 들어가면 다른 하나까지 밀린다
   (사이 피치 11.83 = 줄 900 + 여백 283; `NoteMeasurement.stackingHeight(isNoteEnd:)`,
@@ -1289,7 +1289,13 @@ paraShape와 같은 값**이어야 한다.
   문단을 다시 조판하므로, 구역이 바뀌어 폭이 달라지면 줄 나눔이 달라져 캐시 줄 비례
   환산이 앞 쪽의 마지막 글자와 다른 자리를 가리킨다 (실측: 451 → 200pt에서 7자 유실).
   앞 쪽이 **실제로 그린** 문자 길이를 이월 입력에 실어 그 위치부터 잇는다 — 폭이 그대로면
-  비례 환산 결과가 곧 그 위치라 값이 같다. 경계가 줄 중간에 떨어지면 그 줄의 앞부분을
+  비례 환산 결과가 곧 그 위치라 값이 같다. **각주 모양도 같이 실어 보낸다**
+  (`Input.footnoteShape`): 번호 라벨은 구역 각주 모양의 장식 문자·번호 모양을 따르므로
+  구역이 바뀐 쪽에서 다시 만들면 라벨 길이가 달라져 조판 문자열이 통째로 밀리고, 폭이
+  같아도 `placedLength`가 다른 글자를 가리킨다 (실측: `(1)` → `1`에서 `word45`가 `rd45`로
+  잘림). 예약도 그 모양으로 쟀으므로 같은 값을 써야 예약 ≡ 배치가 성립한다 —
+  `sizeResolver`(R44 #1)와 같은 규약이고, `measure`가 처음 잰 값을 각인해 나른다.
+  구분선 여백·길이는 **쪽**의 몫이라 그대로 현재 구역 모양을 쓴다. 경계가 줄 중간에 떨어지면 그 줄의 앞부분을
   잘라 내되 줄 기하와 앵커는 앞 쪽 폭 기준이라 근사다 (분할 각주는 개체를 안 담으므로
   앵커는 버린다).
   **개체를 담은 각주는 나누지 않는다** (`NoteMeasurement.carriesObjects`): 개체는 앵커
@@ -1298,7 +1304,12 @@ paraShape와 같은 값**이어야 한다.
   나누는지는 **실측이 없어** (코퍼스에 그런 각주 0건) 통째로 옮긴다. 같은 이유로 그런
   각주는 **CT 높이를 그대로 쓴다**: 쪽 몫의 합은 한글이 두 쪽에 나눠 그린 높이인데 우리는
   한 쪽에 통째로 그리므로, 그 합으로 낮추면 블록이 CT 좌표의 개체를 못 담아 다음 각주와
-  겹친다 (실측: 블록 44.16pt에 개체 하단 79.35pt). 여러 문단짜리 각주의 뒤 문단이 0에서 다시 시작하면 그
+  겹친다 (실측: 블록 44.16pt에 개체 하단 79.35pt). **두 판정의 범위는 각주 단위로 같다**
+  (#165 리뷰): 문단 단위로 보면 개체가 뒤 문단에만 있을 때 앞 문단만 캐시 합으로 줄어 그
+  문단의 마지막 글줄 위로 뒤 문단의 그림이 올라온다. 판정은 `measure`가 `noteId`별로 한 번
+  (`hasFloatingObject` 상위집합 술어) 하고, **예약도 같은 범위를 본다** — 개체를 담은
+  각주의 문단은 그 문단에 개체가 없어도 캐시 빠른 길(`measuredFootnoteTextHeight`)로 가지
+  않는다 (가면 예약이 캐시 합, 배치가 CT 높이로 갈린다). 여러 문단짜리 각주의 뒤 문단이 0에서 다시 시작하면 그
   문단부터 통째로 다음 쪽이다. 들어맞음 허용 오차는 부동소수 잡음뿐 — 한글은 1 HWPUNIT
   여유에도 싣는다. 진행 보장: 빈 쪽에도 안 들어가는 분할 지점 없는 각주는 그대로 싣는다
   (코퍼스 0건). 결과: 헌법주석 겹침 366쪽 → **0쪽**, 쪽수 1,030 그대로, 이어짐 쪽 64
